Сделал рефактор кода. Разделил handlers на registration, authorization, main. Переписал сервер, вместо endpoints использовал Flask
