+++ config.py
@@ -5,7 +5,6 @@ from dotenv import load_dotenv
 load_dotenv('.env')
 
 BOT_TOKEN = getenv('TOKEN')
-ADMIN_ID = getenv('ADMIN_ID')
 PG_USER = getenv('PG_USER')
 PG_PASS = str(getenv('PG_PASS'))
 MY_EMAIL = getenv('MY_EMAIL')
