+++ create_url.py
@@ -1,10 +1,3 @@
-from random import choice
-
-
 def create_url(email):
-    chars = 'QWERTYUIOPLKJHGFDSAZXCVBNMqwertyuiopasdfghjklzxcvbnm'
-    param = ''
-    for _ in range(10 + 1):
-        param += choice(chars)
-    url = f'http://localhost:8000/mail?param1={param}&param2={email}'
+    url = f'http://localhost:5000/mail/{email}'
     return url
