+++ message_to_email.py
@@ -9,6 +9,6 @@ def send_password_to_email(email, password):
     smtp_obj.starttls()
     smtp_obj.login(MY_EMAIL, EMAIL_PASSWORD)
     smtp_obj.sendmail(MY_EMAIL, email,
-                      f'Ваш пароль: {password}\nСсылка для подтверждения почты {create_url(email=email)}'.encode(
+                      f'Ваш пароль: {password}\nСсылка для подтверждения почты: {create_url(email=email)}'.encode(
                           encoding='utf-8'))
     smtp_obj.quit()
