+++ requirements.txt
@@ -1,5 +1,4 @@
-endpoints==4.0.5
+Flask==1.1.2
 psycopg2==2.8.6
 pyTelegramBotAPI==3.7.6
 python-dotenv==0.15.0
-
