+++ server.py
@@ -1,0 +1,15 @@
+from flask import Flask
+
+from work_with_db import update_mail_activated
+
+app = Flask(__name__)
+
+
+@app.route("/mail/<email>")
+def mail(email):
+    update_mail_activated(email=email)
+    return f'Email {email} activated'
+
+
+if __name__ == "__main__":
+    app.run()
